Move non-sending tests to a separate class
