+++ com.unity.mobile.notifications/Tests/Runtime/Android/AndroidNotificationSimpleTests.cs
@@ -1,0 +1,65 @@
+using System;
+using NUnit.Framework;
+using Unity.Notifications.Android;
+
+class AndroidNotificationSimpleTests
+{
+	[Test]
+    public void CreateNotificationChannel_NotificationChannelIsCreated()
+    {
+#if !UNITY_EDITOR
+        var testChannelId = "default_test_channel_10";
+        AndroidNotificationCenter.DeleteNotificationChannel(testChannelId);
+        Assert.AreNotEqual("default_test_channel_10", AndroidNotificationCenter.GetNotificationChannel(testChannelId).Id);
+
+        var newChannel = new AndroidNotificationChannel();
+        newChannel.Id = testChannelId;
+        newChannel.Name = "Default Channel";
+        newChannel.Description = "Generic spam";
+
+        var currentChannelCount = AndroidNotificationCenter.GetNotificationChannels().Length;
+        AndroidNotificationCenter.RegisterNotificationChannel(newChannel);
+        currentChannelCount++;
+
+        Assert.AreEqual(currentChannelCount, AndroidNotificationCenter.GetNotificationChannels().Length);
+#endif
+    }
+
+    [Test]
+    public void DeleteNotificationChannel_NotificationChannelIsDeleted()
+    {
+#if !UNITY_EDITOR
+        var ch = new AndroidNotificationChannel();
+        ch.Id = "default_test_channel_0";
+        ch.Name = "Default Channel";
+        ch.Description = "Generic spam";
+        ch.Importance = Importance.Default;
+        AndroidNotificationCenter.RegisterNotificationChannel(ch);
+
+        int numChannels = AndroidNotificationCenter.GetNotificationChannels().Length;
+        AndroidNotificationCenter.DeleteNotificationChannel(ch.Id);
+
+        Assert.AreEqual(numChannels - 1, AndroidNotificationCenter.GetNotificationChannels().Length);
+#endif
+    }
+
+    [Test]
+    public void SetNotificationFireTime_TimeIsConvertedToUnixTimeAndBack()
+    {
+        var n = new AndroidNotification();
+        var fireTime = new DateTime(2018, 5, 24, 12, 41, 30, 122);
+        n.FireTime = new DateTime(2018, 5, 24, 12, 41, 30, 122);
+
+        Assert.AreEqual(fireTime, n.FireTime);
+    }
+
+    [Test]
+    public void SetNotificationRepeatInterval_TimeIsConvertedToUnixTimeAndBack()
+    {
+        var n = new AndroidNotification();
+        var repeatInterval = TimeSpan.FromSeconds(666);
+        n.RepeatInterval = repeatInterval;
+
+        Assert.AreEqual(repeatInterval, n.RepeatInterval);
+    }
+}
