+++ com.unity.mobile.notifications/Tests/Runtime/Android/AndroidNotificationSimpleTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f2d51a3855e994c70bc3396a4f82b55d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
